requirements.txt
Django == latest
django-rest-framework==latest